Create README.md
In this project, the goal was to create an event forecaster. A program that would pull data from both the Ticketmaster API and Open Weather API to display a list of events; that when clicked on, will give general weather information for a given day.

Dates for this project were selected and formatted using Day JS!

The main focus of this project was combining two server-side API'S, which was accomplished relatively quickly.

We decided to use Bulma for our CSS Framework and thius allowed us to implement easy use formatting, dropdowns, and boxes.

For the OpenWeather API, I ended up having to use two separate API call links and used internal logic to determine which path the weather would go.
For the Tickermater Discovery API, I used global variables for the different dates to set up the timeframe for the call and then would append the items to the list on the page using JQuery DOM traversal.

This project for all of us was the first collaborative coding effort, so it was a fantastic learning experience for working in a team, merging conflicts and just overall understanding GitHub's group functionality!